+++ src/i18n/locales/gsw/notifications.ts
@@ -30,6 +30,7 @@ export default {
   game_deleted: 'Spiel abgseit: {{home_team}} gege {{away_team}} am {{date}}',
   game_reinstated: 'Spiel findet wieder statt: {{home_team}} gege {{away_team}} am {{date}}',
   game_result: 'Resultat: {{home_team}} {{home_score}}:{{away_score}} {{away_team}}',
+  game_invite: '{{team}} het di ufbotte: {{matchup}} am {{date}}',
   training_created: 'Neus Training am {{date}}, {{time}} @ {{hall}}',
   training_updated: 'Training aktualisiert am {{date}} @ {{hall}}',
   training_cancelled: 'Training abgseit am {{date}}',